lambda/requirements.txt
boto3
jinja2
pandas
python-dotenv
redshift_connector